agent-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
typing-extensions>=4.0.0
openai==1.50.0
python-dotenv==1.0.1

